diktat-gradle-plugin
### What's done:
* Support for gradle <6.4
* Refactoring
